requirements/dev.txt
-r base.txt
django-debug-toolbar
pytest-django
black
flake8
isort